+++ LR/07/tests/tests.cpp
@@ -1,50 +1,16 @@
-#include "testgateway.h"
+#include "tests.h"
 
-/**
- * @brief Констркутор класса TestGateway, в котором
- * инициализируется объект тестируемого класса Gateway
- * @param parent
- */
-TestGateway::TestGateway(QObject *parent)
-    : QObject(parent)
-{
-    testObj = new Gateway(nullptr);
-    QDomDocument config;
-    QFile file(QDir::currentPath() + "/../../07/tests/getTestSuites/testSuites.xml");
-    if (file.open(QIODevice::ReadOnly)) {
-        if (config.setContent(&file)) {
-            suites = config.documentElement();
-        }
-        file.close();
-    }
-}
-
-/**
- * @brief Тестовая функция отправляет Gateway полностью корректную строчку формата Json
- */
-void TestGateway::testTrueJson()
-{
-    QByteArray inputRightData("{\"messageType\": 1, \"lab\": 7, \"variant\": 1, \"link\": \"https://github.com/leshastern/strategy1\"}");
-
-    try {
-        testObj->validateData(inputRightData);
-    } catch (WrongRequestException error) {
-        qCritical() << error.text() << error.jsonKey() << error.getRejectCode();
-        QFAIL("This case should be passed!");
-    }
-}
 /**
  * @brief Метод, аккумулирующий некорректные Json для далнейшей проверки
  * описания кейсов есть в сценариях (см testsSuites.xlsx)
  */
-void TestGateway::testSuite_data()
+void testSuite_data(QDomElement suites, QRegExp re)
 {
     QTest::addColumn<QByteArray>("testData");
     QTest::addColumn<QString>("description");
     QTest::addColumn<QString>("expected");
 
     if (!suites.isNull()) {
-        QRegExp re("issue-393-[0-9]{1,3}");
         for (QDomElement key = suites.firstChildElement();
              !key.isNull(); key = key.nextSibling().toElement())
         {
@@ -66,7 +32,7 @@ void TestGateway::testSuite_data()
  * @brief Тестовая функция извлекает из заранее подготовленной таблицы
  * некорректные Json и отдает их на проверку методу validateData()
  */
-void TestGateway::testSuite()
+void testSuite(QDomElement suites)
 {
     if (!suites.isNull()) {
         QFETCH(QByteArray, testData);
@@ -74,7 +40,7 @@ void TestGateway::testSuite()
         QFETCH(QString, expected);
 
         try {
-          testObj->validateData(testData);
+        //  testObj->validateData(testData);
           QFAIL("Test worked like input data is valid, but it is invalid (at least it should be)");
         } catch (WrongRequestException error) {
     //        TODO: заменить после задачи https://github.com/Students-of-the-city-of-Kostroma/trpo_automation/issues/51
@@ -88,10 +54,57 @@ void TestGateway::testSuite()
     }
 }
 
+
 /**
- * @brief Почищаем после всех тестов
+ * @brief Констркутор класса TestGateway, в котором
+ * инициализируется объект тестируемого класса Gateway
+ * @param parent
  */
-void TestGateway::cleanupTestCase()
+TestGateway::TestGateway(QObject *parent)
+    : QObject(parent)
 {
-    delete testObj;
+    testObj = new Gateway(nullptr);
+    QDomDocument config;
+    QFile file(QDir::currentPath() + "/../../07/tests/getTestSuites/testSuites.xml");
+    if (file.open(QIODevice::ReadOnly)) {
+        if (config.setContent(&file)) {
+            suites = config.documentElement();
+        }
+        file.close();
+    }
+    QRegExp re("issue-393-[0-9]{1,3}");
+    testSuite_data(suites, re);
+    testSuite(suites);
+}
+
+TestStrategyLab::TestStrategyLab(QObject *parent)
+    : QObject(parent)
+{
+    testObj = new StrategyLab(nullptr);
+    QDomDocument config;
+    QFile file(QDir::currentPath() + "/../../07/tests/getTestSuites/testSuites.xml");
+    if (file.open(QIODevice::ReadOnly)) {
+        if (config.setContent(&file)) {
+            suites = config.documentElement();
+        }
+        file.close();
+    }
+    QRegExp re("issue-173-[0-9]{1,2}");
+    testSuite_data(suites, re);
+    testSuite(suites);
+}
+
+
+int main(int argc,char* argv[])
+{
+    int status = 0;
+   {
+        TestGateway tg;
+        status |= QTest::qExec(&tg, argc, argv);
+   }
+   {
+        TestStrategyLab tsl;
+        status |= QTest::qExec(&tsl, argc, argv);
+   }
+    return status;
 }
